+++ pg_analytics/tests/create.rs
@@ -1,7 +1,6 @@
 mod fixtures;
 
 use fixtures::*;
-use pretty_assertions::assert_eq;
 use rstest::*;
 use sqlx::PgConnection;
